Current work on navi

+++ WalkaRound/WalkaRound-App/src/edu/kit/iti/algo2/pse2013/walkaround/client/model/navigation/output/NaviOutput.java
@@ -1,6 +1,6 @@
 package edu.kit.iti.algo2.pse2013.walkaround.client.model.navigation.output;
 
 public interface NaviOutput {
-	public void deliverOutput(double turnAngle, int distToTurn);
+	public void deliverOutput(double turnAngle, double distToTurn);
 
 }
